+++ Examples/FKKitExamples/FKKitExamples/Examples/BarPresentation/FKBarPresentationExampleViewController.swift
@@ -533,10 +533,3 @@ extension FKBarPresentationExampleViewController: FKBarPresentationDataSource {
     return makePanelContent(for: index, item: item)
   }
 }
-
-private extension Array {
-  subscript(safe index: Int) -> Element? {
-    guard indices.contains(index) else { return nil }
-    return self[index]
-  }
-}