+++ Examples/FKKitExamples/FKKitExamples/Examples/Filter/FKFilterExampleDataProvider.swift
@@ -56,6 +56,9 @@ final class FKFilterExampleMockDataProvider: FKFilterExampleDataProviding, Senda
           .init(id: .init(rawValue: "csw-1-1"), title: "党的二十大"),
           .init(id: .init(rawValue: "csw-1-2"), title: "党务干部"),
           .init(id: .init(rawValue: "csw-1-3"), title: "一把手"),
+          .init(id: .init(rawValue: "csw-1-4"), title: "十九届六中全会"),
+          .init(id: .init(rawValue: "csw-1-5"), title: "十九届五中全会"),
+          .init(id: .init(rawValue: "csw-1-6"), title: "...")
         ]
       ),
       .init(
@@ -66,6 +69,24 @@ final class FKFilterExampleMockDataProvider: FKFilterExampleDataProviding, Senda
           .init(id: .init(rawValue: "csw-2-1"), title: "发行上市"),
           .init(id: .init(rawValue: "csw-2-2"), title: "新三板"),
           .init(id: .init(rawValue: "csw-2-3"), title: "机构业务"),
+          .init(id: .init(rawValue: "csw-2-4"), title: "期货业务"),
+          .init(id: .init(rawValue: "csw-2-5"), title: "稽查处罚"),
+          .init(id: .init(rawValue: "csw-2-6"), title: "...")
+        ]
+      ),
+      .init(
+        id: .init(rawValue: "csw-3"),
+        title: "最新政策法规",
+        selectionMode: .single,
+        items: []
+      ),
+      .init(
+        id: .init(rawValue: "csw-4"),
+        title: "其他",
+        selectionMode: .single,
+        items: [
+          .init(id: .init(rawValue: "csw-4-1"), title: "心理健康与心理..."),
+          .init(id: .init(rawValue: "csw-4-2"), title: "碳达峰碳中和"),
         ]
       ),
     ]
@@ -87,6 +108,11 @@ final class FKFilterExampleMockDataProvider: FKFilterExampleDataProviding, Senda
           .init(id: .init(rawValue: "file-excel"), title: "Excel"),
           .init(id: .init(rawValue: "file-ppt"), title: "PPT"),
           .init(id: .init(rawValue: "file-pdf"), title: "PDF"),
+          .init(id: .init(rawValue: "file-mp3"), title: "MP3"),
+          .init(id: .init(rawValue: "file-mp4"), title: "MP4"),
+          .init(id: .init(rawValue: "file-h5"), title: "H5"),
+          .init(id: .init(rawValue: "file-image"), title: "图片"),
+          .init(id: .init(rawValue: "file-other"), title: "其他"),
         ]
       )
     ]
@@ -117,8 +143,28 @@ final class FKFilterExampleMockDataProvider: FKFilterExampleDataProviding, Senda
         items: [
           .init(id: .init(rawValue: "tag-all"), title: "全部"),
           .init(id: .init(rawValue: "tag-ib"), title: "投资银行"),
+          .init(id: .init(rawValue: "tag-vip"), title: "畅学专享课"),
           .init(id: .init(rawValue: "tag-sell"), title: "证券承销与保荐"),
+          .init(id: .init(rawValue: "tag-investment-research"), title: "投资研究"),
           .init(id: .init(rawValue: "tag-ethics"), title: "职业道德"),
+          .init(id: .init(rawValue: "tag-risk-management"), title: "风险管理"),
+          .init(id: .init(rawValue: "tag-securities-brokerage"), title: "证券经纪"),
+          .init(id: .init(rawValue: "tag-it-management"), title: "信息技术管理"),
+          .init(id: .init(rawValue: "tag-culture"), title: "文化建设"),
+          .init(id: .init(rawValue: "tag-investment-advisory"), title: "投资顾问"),
+          .init(id: .init(rawValue: "tag-compliance"), title: "合规管理"),
+          .init(id: .init(rawValue: "tag-derivatives"), title: "金融衍生品"),
+          .init(id: .init(rawValue: "tag-wealth"), title: "财富管理"),
+          .init(id: .init(rawValue: "tag-party"), title: "党建引领"),
+          .init(id: .init(rawValue: "tag-merger"), title: "并购重组"),
+          .init(id: .init(rawValue: "tag-asset-management"), title: "资产管理"),
+          .init(id: .init(rawValue: "tag-investor-protection"), title: "投资者保护"),
+          .init(id: .init(rawValue: "tag-comprehensive-regulation"), title: "综合性法规"),
+          .init(id: .init(rawValue: "tag-operation"), title: "运营管理"),
+          .init(id: .init(rawValue: "tag-bond"), title: "债券业务"),
+          .init(id: .init(rawValue: "tag-credit"), title: "信用业务"),
+          .init(id: .init(rawValue: "tag-custody"), title: "托管业务"),
+          .init(id: .init(rawValue: "tag-other"), title: "其他"),
         ]
       )
     ]